requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyascii==0.3.2
anyio==4.4.0
appnope==0.1.4
argostranslate==1.9.6
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audio-separator==0.8.0
audio-splitter==0.1.0
audioread==3.0.1
av==12.1.0
Babel==2.15.0
backports.tarfile==1.2.0
bangla==0.0.2
beartype==0.18.5
beautifulsoup4==4.12.3
bitarray==3.4.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
bottle==0.13.4
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contextlib2==21.6.0
contourpy==1.2.1
coqpit==0.0.17
ctranslate2==4.3.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
debugpy==1.8.2
decorator==4.4.2
deep-translator==1.11.4
deepspeed==0.9.5
diffq==0.2.4
dnspython==2.6.1
docopt==0.6.2
docutils==0.21.2
dtw-python==1.5.1
einops==0.8.0
email_validator==2.1.1
encodec==0.1.1
eng-to-ipa==0.0.2
exceptiongroup==1.2.1
executing==2.0.1
fairseq==0.12.2
fastapi==0.111.0
fastapi-cli==0.0.4
faster-whisper==1.0.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.15.1
Flask==3.0.3
flask-cors==6.0.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
g2pkk==0.1.2
gast==0.6.0
google-pasta==0.2.0
googletrans==3.0.0
gradio==3.48.0
gradio_client==0.6.1
grpcio==1.64.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hangul-romanize==0.1.0
hjson==3.1.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.23.4
humanfriendly==10.0
hydra-core==1.0.7
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==6.2.0
ipykernel==6.29.4
ipython==8.26.0
itsdangerous==2.2.0
jamo==0.4.1
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.7.0
keyring==25.2.1
kiwisolver==1.4.5
langcodes==3.4.0
-e git+https://github.com/FlipSuite/LangFlip.git@1fcc0ab546122c93d315b6267b3a8df42ab0603f#egg=langflip
langid==1.1.6
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.9.2
llvmlite==0.43.0
lxml==5.4.0
macholib==1.16.3
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mutagen==1.47.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==2.8.8
nh3==0.2.17
ninja==1.11.1.1
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
omegaconf==2.0.6
onnx==1.16.1
onnx2torch==1.5.14
onnxruntime==1.14.0
openai-whisper==20231117
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.1
pandas==1.5.3
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
pkginfo==1.11.1
platformdirs==4.2.2
pooch==1.8.2
portalocker==3.1.1
preshed==3.0.9
proces==0.1.7
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==4.25.3
proxy-tools==0.1.0
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==1.10.18
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pynndescent==0.5.13
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyobjc-framework-Security==11.1
pyobjc-framework-WebKit==11.1
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywebview==6.0
PyYAML==6.0.1
pyzmq==26.0.3
readme_renderer==43.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
rfc3986==1.5.0
rich==13.7.1
rotary-embedding-torch==0.6.2
rpds-py==0.18.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
sacrebleu==2.5.1
sacremoses==0.0.53
safetensors==0.4.3
samplerate==0.1.0
scikit-learn==1.5.0
scipy==1.11.4
selenium==4.22.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
srsly==2.4.8
stack-data==0.6.3
stanza==1.1.1
starlette==0.37.2
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12.1
tabulate==0.9.0
tacotron2==22.12.28
tacotron2-model==0.2.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.5.0
textwrap3==0.9.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
trainer==0.0.36
traitlets==5.14.3
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
TTS==0.22.0
twine==5.1.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
waveglow==22.12.28
wavmark==0.0.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
wget==3.2
whisper-timestamped==1.15.4
wrapt==1.16.0
wsproto==1.2.0
xgboost==3.0.2
yarl==1.9.4
zipp==3.19.2
